WIP - v4 - ctrl+z shortcut, restore one delete entry at once.

+++ src/MainWindow.vala
@@ -11,8 +11,8 @@ namespace HostsManager {
 		private const ActionEntry[] ACTION_ENTRIES = {
 			{ "focus-search-bar", focus_search_bar },
 			{ "host-row-add", host_row_add },
-			{ "host-row-delete", host_row_delete},
-			{ "host-row-undo", signal_toast_undo_button_clicked_handler },
+			{ "host-row-delete", host_row_delete },
+			{ "host-row-undo", signal_shortcut_undo_triggered },
 			{ "restore-fron-backup", restore_from_backup },
 			{ "show-about", show_about },
 			{ "app-quit", app_quit },
@@ -259,7 +259,7 @@ namespace HostsManager {
 			GLib.ListStore host_rows_to_delete = new GLib.ListStore (typeof (Models.HostRow));
 
 			if (!iter.init_first (this.hosts_multi_selection.get_selection (), out position)) {
-			return;
+				return;
 			}
 
 			do {
@@ -293,12 +293,12 @@ namespace HostsManager {
 			this.toast_overlay.add_toast (this.toast_undo);
 		}
 
-		[GtkCallback]
-		private void signal_toast_undo_button_clicked_handler () {
+		private void host_row_delete_undo (bool from_shortcut = false) {
 
 			if (this.hosts_list_undo_store.get_n_items () > 0) {
 
-				for (int position = 0; position <= this.hosts_list_undo_store.n_items; position++) {
+				int position;
+				for (position = 0; position <= this.hosts_list_undo_store.n_items; position++) {
 
 					Models.HostRow host_row = this.hosts_list_undo_store.get_item (position) as Models.HostRow;
 					if (host_row != null) {
@@ -306,16 +306,41 @@ namespace HostsManager {
 						debug ("Restoring host \"%s\", IP address \"%s\", previous position: %u", host_row.hostname, host_row.ip_address, host_row.previous_position);
 						this.hosts_list_store.insert (host_row.previous_position, host_row);
 						this.hosts_file_service.restore (host_row, false);
+
+						if (from_shortcut == true) {
+
+							break;
+						}
 					}
 				}
 
-				this.hosts_list_undo_store.remove_all ();
+				if (from_shortcut == false) {
+
+					this.hosts_list_undo_store.remove_all ();
+				} else {
+
+					this.hosts_list_undo_store.remove (position);
+				}
+
 				this.hosts_file_service.save_file ();
 			} else {
 
 				this.toast.title = _("No deleted entries to restore.");
 				this.toast_overlay.add_toast (this.toast);
 			}
+		}
+
+		private void signal_shortcut_undo_triggered () {
+
+			// will undo only one row at once.
+			this.host_row_delete_undo (true);
+		}
+
+		[GtkCallback]
+		private void signal_toast_undo_button_clicked_handler () {
+
+			// will undo all deleted rows.
+			this.host_row_delete_undo (false);
 		}
 
 		private void restore_from_backup () {
@@ -425,7 +450,7 @@ namespace HostsManager {
 				editable_cell.field_type = Widgets.EditableCell.FieldType.HOSTNAME;
 				editable_cell.editable_label.set_text (host_row.hostname);
 				editable_cell.host_row = host_row;
-				//editable_cell.initDragAndDrop ();
+				// editable_cell.initDragAndDrop ();
 			}
 		}
 
@@ -440,41 +465,41 @@ namespace HostsManager {
 		}
 
 		// Drag & Drop
-		//  public void handle_drop (uint drag_item_position, uint drop_item_position) {
+		// public void handle_drop (uint drag_item_position, uint drop_item_position) {
 
 
-		//  	var iter = Gtk.BitsetIter ();
-		//  	uint position;
-		//  	uint initial_position = 0;
-		//  	bool initial_position_init = false;
-		//  	Object[] host_rows_to_move = {};
+		// var iter = Gtk.BitsetIter ();
+		// uint position;
+		// uint initial_position = 0;
+		// bool initial_position_init = false;
+		// Object[] host_rows_to_move = {};
 
-		//  	debug ("drag_item_position: %u | drop_item_position: %u", drag_item_position, drop_item_position);
-		//  	// if (!iter.init_first (this.hosts_multi_selection.get_selection (), out position)) {
-		//  	// debug ("par la");
-		//  	// return;
-		//  	// }
+		// debug ("drag_item_position: %u | drop_item_position: %u", drag_item_position, drop_item_position);
+		//// if (!iter.init_first (this.hosts_multi_selection.get_selection (), out position)) {
+		//// debug ("par la");
+		//// return;
+		//// }
 
-		//  	iter.init_first (this.hosts_multi_selection.get_selection (), out position);
-		//  	do {
+		// iter.init_first (this.hosts_multi_selection.get_selection (), out position);
+		// do {
 
-		//  		debug ("M - par la");
-		//  		Models.HostRow host_row = this.hosts_list_store.get_item (position) as Models.HostRow;
-		//  		if (host_row != null && host_row.row_type == Models.HostRow.RowType.HOST) {
+		// debug ("M - par la");
+		// Models.HostRow host_row = this.hosts_list_store.get_item (position) as Models.HostRow;
+		// if (host_row != null && host_row.row_type == Models.HostRow.RowType.HOST) {
 
-		//  			if (initial_position_init == false) {
+		// if (initial_position_init == false) {
 
-		//  				initial_position_init = true;
-		//  				initial_position = position;
-		//  			}
+		// initial_position_init = true;
+		// initial_position = position;
+		// }
 
-		//  			host_rows_to_move += host_row;
-		//  		}
-		//  	} while (iter.next (out position));
+		// host_rows_to_move += host_row;
+		// }
+		// } while (iter.next (out position));
 
-		//  	this.host_row_delete (true);
-		//  	debug ("drop_item_position: %u | num host_rows_to_move:  %u", drop_item_position, host_rows_to_move.length);
-		//  	this.hosts_list_store.splice (drop_item_position, 0, host_rows_to_move);
-		//  }
+		// this.host_row_delete (true);
+		// debug ("drop_item_position: %u | num host_rows_to_move:  %u", drop_item_position, host_rows_to_move.length);
+		// this.hosts_list_store.splice (drop_item_position, 0, host_rows_to_move);
+		// }
 	}
 }
